Cleaned up the server.js accesses with "status"
Can send for a username, stat=0 to begin the querying process. In order
to compute stats, send the same username with stat=1 a few seconds
later. If any other combination of username (that is different from the
first one) and stat is received, code scream foul. test.html provided
for  testing and how to make GET() requests